refactor: move payment type back to types
